tests passing for ibis
